% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greeter.R
\name{greeter}
\alias{greeter}
\title{Title}
\usage{
greeter(name)
}
\arguments{
\item{name}{A character string indicating the name of the 2nd party}
}
\value{
A statement that greets a person
}
\description{
Title
}
\examples{
greeter("Shel")
}
